Separate librustc_data_structures module
